Moved files into the documents directory for clearer splitting
