+++ studio/schemas/documents/author.js
