+++ studio/schemas/documents/category.js
